Create temporary directory during website generation
Using a temporary directory makes it possible to only override the
generated result in case the full generation was successful.

Currently, the website generation uses assert and panic all over the
place. In order to catch the errors at runtime (which is necessary for
a gentle exit after a failed generation), reorganise all their
occurences into error types.

+++ src/builder.rs
@@ -4,6 +4,7 @@ use std::fs::File;
 use std::io::Error as IOError;
 use std::path::{Path, PathBuf};
 
+mod fileutil;
 mod rendering;
 mod rss;
 mod serialize;
@@ -12,12 +13,13 @@ mod website;
 
 use serialize::LayoutInfo;
 use theme::{TemplateType, Theme, ThemeError};
-use website::{BlogElement, LoadError, OrgFile, Project, Website};
+use website::{BlogElement, OrgFile, Project, Website, WebsiteError};
 
 #[derive(Debug)]
 pub enum InitError {
     Theme(ThemeError),
-    Website(LoadError),
+    Website(WebsiteError),
+    IO(IOError),
 }
 
 impl From<ThemeError> for InitError {
@@ -26,18 +28,26 @@ impl From<ThemeError> for InitError {
     }
 }
 
-impl From<LoadError> for InitError {
-    fn from(err: LoadError) -> Self {
+impl From<WebsiteError> for InitError {
+    fn from(err: WebsiteError) -> Self {
         Self::Website(err)
+    }
+}
+
+impl From<IOError> for InitError {
+    fn from(err: IOError) -> Self {
+        Self::IO(err)
     }
 }
 
 #[derive(Debug)]
 pub enum RenderError {
     Theme(theme::RenderError),
-    HTML(rendering::HTMLExportError),
+    HTML(rendering::SerializationError),
     IO(IOError),
     RSS(rss::Error),
+    FileNotFound(String),
+    InvalidImageDependency { file: String, dependency: String },
 }
 
 impl From<theme::RenderError> for RenderError {
@@ -46,8 +56,8 @@ impl From<theme::RenderError> for RenderError {
     }
 }
 
-impl From<rendering::HTMLExportError> for RenderError {
-    fn from(err: rendering::HTMLExportError) -> Self {
+impl From<rendering::SerializationError> for RenderError {
+    fn from(err: rendering::SerializationError) -> Self {
         Self::HTML(err)
     }
 }
@@ -67,16 +77,18 @@ impl From<rss::Error> for RenderError {
 pub struct Builder<'a> {
     theme: Theme<'a>,
     website: Website,
+    temp_dir: PathBuf,
+    output_path: &'a str,
 }
 
 pub trait Mode: Sized {
-    fn create(output_path: &str) -> Self;
+    fn create(builder: &Builder) -> Self;
     fn base_url(&self) -> String;
 
-    fn include_page(&self, page: &OrgFile) -> bool;
-    fn include_post(&self, post: &OrgFile) -> bool;
-    fn include_project(&self, project: &Project) -> bool;
-    fn include_rss(&self) -> bool;
+    fn include_page(page: &OrgFile) -> Result<bool, website::OrgFileError>;
+    fn include_post(post: &OrgFile) -> Result<bool, website::OrgFileError>;
+    fn include_project(project: &Project) -> Result<bool, website::ProjectError>;
+    fn include_rss() -> bool;
 }
 
 pub struct ReleaseMode {}
@@ -85,7 +97,7 @@ pub struct PreviewMode {
 }
 
 impl Mode for ReleaseMode {
-    fn create(_output_path: &str) -> Self {
+    fn create(_: &Builder) -> Self {
         Self {}
     }
 
@@ -93,174 +105,164 @@ impl Mode for ReleaseMode {
         String::from("https://jhuwald.com")
     }
 
-    fn include_page(&self, page: &OrgFile) -> bool {
-        page.published.is_some()
+    fn include_page(page: &OrgFile) -> Result<bool, website::OrgFileError> {
+        Ok(page.published.is_some())
     }
 
-    fn include_post(&self, post: &OrgFile) -> bool {
+    fn include_post(post: &OrgFile) -> Result<bool, website::OrgFileError> {
         if post.published.is_none() {
-            return false;
+            return Ok(false);
         }
-        assert!(
-            post.from_preamble("summary").is_some() || post.post_type == website::PostType::Mini,
-            "published post {:?} is missing a summary",
-            post.path
-        );
-        assert!(
-            post.from_preamble("subtitle").is_some() || post.post_type == website::PostType::Mini,
-            "published post {:?} is missing a summary",
-            post.path
-        );
-        return true;
+
+        if post.post_type == website::PostType::Mini {
+            return Ok(true);
+        }
+
+        if post.from_preamble("summary").is_none() {
+            return Err(website::OrgFileError::MissingRequiredField("summary"));
+        }
+
+        if post.from_preamble("subtitle").is_none() {
+            return Err(website::OrgFileError::MissingRequiredField("subtitle"));
+        }
+
+        return Ok(true);
     }
 
-    fn include_project(&self, project: &Project) -> bool {
-        project.published()
+    fn include_project(project: &Project) -> Result<bool, website::ProjectError> {
+        Ok(project.published())
     }
 
-    fn include_rss(&self) -> bool {
+    fn include_rss() -> bool {
         true
     }
 }
 
 impl Mode for PreviewMode {
-    fn create(output_path: &str) -> Self {
-        let path = PathBuf::from(output_path).canonicalize().unwrap();
-        let path = path.to_str().unwrap().to_string();
-
-        Self { path }
+    fn create(builder: &Builder) -> Self {
+        let path = PathBuf::from(builder.output_path).canonicalize().unwrap();
+        Self {
+            path: path.to_str().unwrap().to_string(),
+        }
     }
 
     fn base_url(&self) -> String {
         self.path.clone()
     }
 
-    fn include_page(&self, _page: &OrgFile) -> bool {
-        true
+    fn include_page(_: &OrgFile) -> Result<bool, website::OrgFileError> {
+        Ok(true)
     }
 
-    fn include_post(&self, post: &OrgFile) -> bool {
-        if post.published.is_some() {
-            if post.from_preamble("summary").is_none() && post.post_type != website::PostType::Mini
-            {
-                println!("published post {:?} is missing a summary", post.path);
-            }
-            if post.from_preamble("subtitle").is_none() && post.post_type != website::PostType::Mini
-            {
-                println!("published post {:?} is missing a subtitle", post.path);
-            }
+    fn include_post(post: &OrgFile) -> Result<bool, website::OrgFileError> {
+        if let Err(err) = ReleaseMode::include_post(post) {
+            println!("Warning: {:?} in {:?}", err, post.path);
         }
-        true
+        Ok(true)
     }
 
-    fn include_project(&self, _project: &Project) -> bool {
-        true
+    fn include_project(_: &Project) -> Result<bool, website::ProjectError> {
+        Ok(true)
     }
 
-    fn include_rss(&self) -> bool {
+    fn include_rss() -> bool {
         false
     }
 }
 
-impl Builder<'_> {
-    pub fn new(website_path: &str, theme_path: &str) -> Result<Self, InitError> {
+impl<'a> Builder<'a> {
+    pub fn new<TMode: Mode>(
+        website_path: &str,
+        theme_path: &str,
+        output_path: &'a str,
+    ) -> Result<Self, InitError> {
+        let theme = Theme::load(theme_path)?;
+        let website = Website::load::<TMode>(website_path)?;
+
+        let mut temp_dir = std::env::temp_dir();
+        temp_dir.push("silas-generated-output");
+
+        assert!(
+            !temp_dir.is_dir(),
+            "Silas: temporary directory already exists, aborting generation."
+        );
+        fs::create_dir(&temp_dir)?;
+
         Ok(Builder {
-            theme: Theme::load(theme_path)?,
-            website: Website::load(website_path)?,
+            theme,
+            website,
+            temp_dir,
+            output_path,
         })
     }
 
-    fn clear_output_path(output_path: &str, overwrite: bool) -> Result<(), RenderError> {
-        let metadata = fs::metadata(output_path);
-        if metadata.is_ok() {
-            if !overwrite {
-                panic!("Target path '{}' is non-empty", output_path);
-            }
-
-            let metadata = metadata.unwrap();
-            if metadata.is_file() {
-                fs::remove_file(output_path)?;
-            } else {
-                fs::remove_dir_all(output_path)?;
-            }
-            println!("Cleared previous result");
+    pub fn copy_generated_files(&self) -> Result<(), IOError> {
+        if fs::metadata(self.output_path).is_ok() {
+            // clear previous result
+            fs::remove_dir_all(self.output_path)?;
         }
-        Ok(())
+
+        let res = fileutil::copy_folder_recursively(&self.temp_dir, self.output_path);
+        self.clear_generated_files();
+        res
     }
 
-    pub fn generate_single_file<TMode: Mode>(
-        &self,
-        file_path: &str,
-        output_path: &str,
-        overwrite_existing: bool,
-    ) -> Result<(), RenderError> {
-        Self::clear_output_path(output_path, overwrite_existing)?;
-        let mode = TMode::create(output_path);
+    pub fn clear_generated_files(&self) {
+        fs::remove_dir_all(&self.temp_dir).unwrap();
+    }
+
+    pub fn generate_single_file<TMode: Mode>(&self, file_path: &str) -> Result<(), RenderError> {
+        let mode = TMode::create(self);
         let layout = LayoutInfo::new(&self.website, &mode);
 
-        let post = self.website.resolve_path(Path::new(file_path));
-        assert!(post.is_some(), "File `{}` not found", file_path);
-
-        let ser = post.unwrap().serialize(&self.website, &mode, &layout)?;
-        let file = File::create(output_path)?;
-        self.render_element(file, TemplateType::Post, &ser)?;
-
-        Ok(())
+        match self.website.resolve_path(Path::new(file_path)) {
+            None => Err(RenderError::FileNotFound(file_path.to_string())),
+            Some(post) => {
+                let ser = post.serialize(&self.website, &mode, &layout)?;
+                let file =
+                    File::create(self.temp_dir.to_str().unwrap().to_string() + "/index.html")?;
+                self.render_element(file, TemplateType::Post, &ser)?;
+                Ok(())
+            }
+        }
     }
 
-    pub fn generate<TMode: Mode>(
-        &self,
-        output_path: &str,
-        overwrite_existing: bool,
-    ) -> Result<(), RenderError> {
-        Self::clear_output_path(output_path, overwrite_existing)?;
+    pub fn generate<TMode: Mode>(&self) -> Result<(), RenderError> {
+        let mode = TMode::create(self);
 
-        fs::create_dir(output_path)?;
-        self.theme.copy_files(output_path)?;
+        self.theme.copy_files(self.temp_dir.to_str().unwrap())?;
 
-        let mode = TMode::create(output_path);
         let layout = LayoutInfo::new(&self.website, &mode);
         let mut rss = rss::RSSBuilder::new(&self.website, &mode);
 
         let mut ser = self.website.serialize(&mode, &layout)?;
-        let file = self.prepare_file(&self.website, output_path, &mut ser.folder_out)?;
+        let file = self.prepare_file(&self.website, &mut ser.folder_out)?;
         self.render_element(file, TemplateType::Page, &ser)?;
         rss.insert_file(&ser);
 
         for page in self.website.pages.values() {
-            if !mode.include_page(&page) {
-                continue;
-            }
             let mut ser = page.serialize(&self.website, &mode, &layout)?;
             rss.insert_file(&ser);
-            let file = self.prepare_file(page, output_path, &mut ser.folder_out)?;
+            let file = self.prepare_file(page, &mut ser.folder_out)?;
             self.render_element(file, TemplateType::Page, &ser)?;
         }
 
         for project in self.website.projects.values() {
-            if !mode.include_project(project) {
-                continue;
-            }
-
             let mut ser = project.serialize(&self.website, &mode, &layout)?;
             rss.start_project(project.id(), &ser);
-            let file = self.prepare_file(project, output_path, &mut ser.folder_out)?;
+            let file = self.prepare_file(project, &mut ser.folder_out)?;
             self.render_element(file, TemplateType::Project(project.project_type), &ser)?;
 
             for post in project.posts.values() {
-                if !mode.include_post(&post) {
-                    continue;
-                }
                 let mut ser = post.serialize(&self.website, &mode, &layout)?;
                 rss.insert_file(&ser);
-                let file = self.prepare_file(post, output_path, &mut ser.folder_out)?;
+                let file = self.prepare_file(post, &mut ser.folder_out)?;
                 self.render_element(file, TemplateType::Post, &ser)?;
             }
             rss.finish_project();
         }
 
-        rss.write_feeds(output_path)?;
-
+        rss.write_feeds(self.temp_dir.to_str().unwrap())?;
         Ok(())
     }
 
@@ -273,12 +275,12 @@ impl Builder<'_> {
         for img in elem.image_deps.iter() {
             let mut path = PathBuf::from(&elem.folder_in);
             path.push(img);
-            assert!(
-                path.is_file(),
-                "Image dependency {:?} for {} doesn't exist",
-                path,
-                elem.url
-            );
+            if !path.is_file() {
+                return Err(RenderError::InvalidImageDependency {
+                    file: elem.url.clone(),
+                    dependency: path.to_str().unwrap().to_string(),
+                });
+            }
             fs::copy(
                 elem.folder_in.clone() + "/" + img,
                 elem.folder_out.clone() + "/" + img,
@@ -291,10 +293,10 @@ impl Builder<'_> {
     fn prepare_file<T: BlogElement>(
         &self,
         elem: &T,
-        output_path: &str,
         folder_out: &mut String,
     ) -> Result<File, IOError> {
-        let filename = elem.url(&self.website, output_path.to_string());
+        let filename = self.temp_dir.to_str().unwrap().to_string();
+        let filename = elem.url(&self.website, filename);
         fs::create_dir_all(&filename)?;
         *folder_out = filename.clone();
         let filename = filename + "/index.html";
